+++ BTCPayServer/Forms/IFormComponentProvider.cs
@@ -1,0 +1,8 @@
+﻿using BTCPayServer.Abstractions.Form;
+
+namespace BTCPayServer.Forms;
+
+public interface IFormComponentProvider
+{
+    public string CanHandle(Field field);
+}
